requirements.txt
aiobotocore==2.26.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aioitertools==0.13.0
aiosignal==1.4.0
alembic==1.17.2
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.0
appdirs==1.4.4
asyncssh==2.22.0
atpublic==7.0.0
attrs==25.4.0
awscli==1.44.11
billiard==4.2.4
blinker==1.9.0
boto3==1.41.5
botocore==1.42.21
cachetools==6.2.4
celery==5.6.1
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpickle==3.1.2
colorama==0.4.6
configobj==5.0.9
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
databricks-sdk==0.76.0
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
docutils==0.19
dpath==2.2.0
dulwich==0.25.0
dvc==3.65.0
dvc-data==3.17.0
dvc-http==2.32.0
dvc-objects==5.2.0
dvc-render==1.0.2
dvc-s3==3.2.2
dvc-studio-client==0.22.0
dvc-task==0.40.2
entrypoints==0.4
fastapi==0.128.0
filelock==3.20.1
Flask==3.1.2
flask-cors==6.0.2
flatten-dict==0.4.2
flufl.lock==8.2.0
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.12.0
funcy==2.0
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.45.0
grandalf==0.8
graphene==3.4.3
graphql-core==3.2.7
graphql-relay==3.2.0
gto==1.9.0
gunicorn==23.0.0
h11==0.16.0
huey==2.5.5
hydra-core==1.3.2
idna==3.11
imbalanced-learn==0.14.1
importlib_metadata==8.7.1
iterative-telemetry==0.0.10
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.3
kiwisolver==1.4.9
kombu==5.6.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mlflow==3.8.1
mlflow-skinny==3.8.1
mlflow-tracing==3.8.1
multidict==6.7.0
networkx==3.6.1
nltk==3.9.2
numpy==2.4.0
omegaconf==2.3.0
opentelemetry-api==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.5
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
pillow==12.1.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.2
psutil==7.2.1
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydot==4.0.1
pygit2==1.19.1
Pygments==2.19.2
pygtrie==2.5.0
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
rich==14.2.0
rsa==4.7.2
ruamel.yaml==0.19.0
ruamel.yaml.clibz==0.3.5
s3fs==2025.12.0
s3transfer==0.16.0
scikit-learn==1.8.0
scipy==1.16.3
scmrepo==3.6.0
semver==3.0.4
setuptools==80.9.0
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.8.0
six==1.17.0
sklearn-compat==0.1.5
smmap==5.0.2
SQLAlchemy==2.0.45
sqlparse==0.5.5
sqltrie==0.11.2
starlette==0.50.0
tabulate==0.9.0
threadpoolctl==3.6.0
tomlkit==0.13.3
tqdm==4.67.1
typer==0.21.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.2
uvicorn==0.40.0
vine==5.1.0
voluptuous==0.16.0
wcwidth==0.2.14
Werkzeug==3.1.4
wrapt==1.17.3
xgboost==2.0.3
yarl==1.22.0
zc.lockfile==4.0
zipp==3.23.0
